backend: category api - view separated
